refactor finding features into its own lib and expose it to the public api
